Rename "api" to "services"
