+++ packages/app/src/components/Home/HomePage.tsx
@@ -7,7 +7,7 @@ import Sources from './Source/Sources'
 import HomeHeader from './Header'
 import { Resources } from './Resources/Resources'
 import { Navigate } from 'react-router-dom'
-import { useGetHealth } from '../../api/health.api'
+import { useGetHealth } from '../../services/health.service'
 
 const Body = styled('div')(({ theme }) => ({
   display: 'flex',
