+++ packages/app/src/services/query.service.ts
@@ -1,6 +1,7 @@
 import { BASE_API_URL } from '.'
 import { apiRequest } from '../utils/request'
 
+// do not use this directly in a component, make a react query for it instead like useExportPulses
 const sqlQueryFn = (query: string) =>
   apiRequest({
     url: `${BASE_API_URL}/localdb/query`,
