+++ packages/app/src/services/version.service.ts
